% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listObjects.R
\name{listObjects}
\alias{listObjects}
\title{List objects in a binary R data file}
\usage{
listObjects(fileName)
}
\arguments{
\item{fileName}{Character string specifying the name of the binary R data file.}
}
\value{
A data frame with columns: "ObjectName", "ObjectType", "ObjectDimensions".
}
\description{
This function loads a binary R data file and returns a data frame
with information about the contained objects.
}
\examples{
X<- matrix(runif(100), nrow=20)
Y<- data.frame(Y1=rnorm(10), Y2=rpois(10, lambda=1), sex=sample(c(0,1), 10, replace=TRUE))
Z=c(X, Y)
save(X,Y,Z, file="binData.Rda")
listObjects(fileName = "binData.Rda")

}
